backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
asyncpg
python-multipart